requirements.txt
beautifulsoup4==4.8.2
requests==2.22.0
pillow==7.0.0
pycrypto==2.6.1
pycryptodome=3.8.2